# 02-CareBridge-AI/prototype/lib/api-zod/src/generated/types/careBridgeExplanationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareBridgeExplanationResultCode } from './careBridgeExplanationResultCode';
import type { CareBridgeExplanationResultProvider } from './careBridgeExplanationResultProvider';
import type { CareBridgeExplanationResultStatus } from './careBridgeExplanationResultStatus';
import type { CareBridgeSourceEvidence } from './careBridgeSourceEvidence';

export interface CareBridgeExplanationResult {
  status: CareBridgeExplanationResultStatus;
  code: CareBridgeExplanationResultCode;
  message: string;
  /** @nullable */
  explanation: string | null;
  source: CareBridgeSourceEvidence | null;
  retryable: boolean;
  /**
     * @minimum 0
     * @maximum 2
     */
  attempts: number;
  provider: CareBridgeExplanationResultProvider;
  liveModelVerified: boolean;
  handoffNotice: string;
}