# 02-CareBridge-AI/prototype/lib/api-zod/src/generated/types/careBridgeSourceEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareBridgeSourceEvidenceSourceType } from './careBridgeSourceEvidenceSourceType';

export interface CareBridgeSourceEvidence {
  patientName: string;
  documentTitle: string;
  sectionTitle: string;
  version: string;
  simulatedApprovalDate: Date;
  exactInstruction: string;
  sourceType: CareBridgeSourceEvidenceSourceType;
}