# 02-CareBridge-AI/prototype/lib/api-zod/src/generated/types/careBridgeTeachBackHistory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareBridgeTeachBackResult } from './careBridgeTeachBackResult';

export interface CareBridgeTeachBackHistory {
  activity: CareBridgeTeachBackResult | null;
  retentionNotice: string;
}