# 02-CareBridge-AI/prototype/lib/api-zod/src/generated/types/careBridgeTeachBackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareBridgeTeachBackInput {
  /**
     * @minLength 8
     * @maxLength 80
     */
  sessionId: string;
  /**
     * @minLength 8
     * @maxLength 80
     */
  idempotencyKey: string;
  /** @maxLength 20 */
  patientId: string;
  /** @maxLength 40 */
  sectionId: string;
  /** @maxLength 40 */
  sourceVersion: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  answer: string;
}