# 02-CareBridge-AI/prototype/lib/api-zod/src/generated/types/careBridgeTeachBackResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareBridgeTeachBackAuditEvent } from './careBridgeTeachBackAuditEvent';
import type { CareBridgeTeachBackResultAssessmentSource } from './careBridgeTeachBackResultAssessmentSource';
import type { CareBridgeTeachBackResultOutcome } from './careBridgeTeachBackResultOutcome';

export interface CareBridgeTeachBackResult {
  activityId: string;
  patientId: string;
  sectionId: string;
  sourceVersion: string;
  originalInstruction: string;
  outcome: CareBridgeTeachBackResultOutcome;
  feedback: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  comprehensionAttempts: number;
  retryAvailable: boolean;
  handoffAvailable: boolean;
  /**
     * @minimum 0
     * @maximum 2
     */
  providerAttempts: number;
  assessmentSource: CareBridgeTeachBackResultAssessmentSource;
  liveModelVerified: boolean;
  updatedAt: Date;
  audit: CareBridgeTeachBackAuditEvent[];
  retentionNotice: string;
}